[builtins] Handle entry access side effects from Map/WeakMap constructor
While adding initial entries to a Map, accessing the key or value
of an entry may call user code that may break fast path
assumptions (initial entries' elements kind or length).

This CL adds handling for this case by detecting for broken
assumptions and falling back to a slower (but safer) TFS
continuation.